requirements.txt
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
python-docx==1.1.0
matplotlib==3.8.2
Pillow==10.2.0
